elections of 1936 California part 1
These focuses are being added in to show how the elctions of 1936 with lead up to. histroically its merriam winning again, and depending on who you align with the most(corpatisit plutocrats or progressive labbor) will decide how the civil war will start. if you go the syndie path, the military will overthrow the government and a civil war ensues betwee nPatton the syndie leader